Split main.go into 2 executables to have an immediate-run-command binary. Update makefile
